int substituídos por size_t

+++ include/cemdutil/array_list.h
@@ -9,11 +9,11 @@
 #define ARRAY_LIST_DEFAULT_STRATEGY_REALLOCATED ARRAY_LIST_HALF_STRATEGY_REALLOCATED // Estratégia padrão de realocação de memória
 
 #define ARRAY_LIST_TYPEDEF(varname, type) \
-    void* __##varname##_array_list_pointer_get(void* pointer, int index) {\
+    void* __##varname##_array_list_pointer_get(void* pointer, size_t index) {\
         type##* p = (type*) pointer;\
         return &p[index];\
     }\
-    void __##varname##_array_list_pointer_set(void* pointer, int index, void* value) {\
+    void __##varname##_array_list_pointer_set(void* pointer, size_t index, void* value) {\
         type##* p = (type*) pointer;\
         type##* v = (type*) value;\
         p[index] = *v;\
@@ -28,8 +28,8 @@
 
 typedef struct __st_array_list_type {
     size_t sizeof_unit;
-    void* (*get)(void* pointer, int index);
-    void (*set)(void* pointer, int index, void* value);
+    void* (*get)(void* pointer, size_t index);
+    void (*set)(void* pointer, size_t index, void* value);
 } ArrayListType;
 
 /**
@@ -40,7 +40,7 @@ typedef struct __st_array_list_type {
  * @return novo espaço a ser realocado (se não for o suficiente, a função será
  *         chamada novamente)
  */
-typedef int ArrayListReallocateStrategy(int length_allocated, int lenght);
+typedef size_t ArrayListReallocateStrategy(size_t length_allocated, size_t lenght);
 
 /**
  * Cria uma estratégia de realocar apenas o espaço necessário pelo
@@ -51,7 +51,7 @@ typedef int ArrayListReallocateStrategy(int length_allocated, int lenght);
  * @return novo espaço a ser realocado (se não for o suficiente, a função será
  *         chamada novamente)
  */
-int ARRAY_LIST_STRICT_STRATEGY_REALLOCATED(int length_allocated, int size);
+size_t ARRAY_LIST_STRICT_STRATEGY_REALLOCATED(size_t length_allocated, size_t size);
 
 /**
  * É a estratégia de realocar adicionando a metade do tamanho do array
@@ -61,7 +61,7 @@ int ARRAY_LIST_STRICT_STRATEGY_REALLOCATED(int length_allocated, int size);
  * @return novo espaço a ser realocado (se não for o suficiente, a função será
  *         chamada novamente)
  */
-int ARRAY_LIST_HALF_STRATEGY_REALLOCATED(int length_allocated, int size);
+size_t ARRAY_LIST_HALF_STRATEGY_REALLOCATED(size_t length_allocated, size_t size);
 
 /**
  * É a estratégia de realocar o dobro do que foi alocado anteriormente
@@ -71,7 +71,7 @@ int ARRAY_LIST_HALF_STRATEGY_REALLOCATED(int length_allocated, int size);
  * @return novo espaço a ser realocado (se não for o suficiente, a função será
  *         chamada novamente)
  */
-int ARRAY_LIST_DOUBLE_STRATEGY_REALLOCATED(int length_allocated, int size);
+size_t ARRAY_LIST_DOUBLE_STRATEGY_REALLOCATED(size_t length_allocated, size_t size);
 
 /**
  * Struct que representa uma instância do ArrayList
@@ -79,13 +79,13 @@ int ARRAY_LIST_DOUBLE_STRATEGY_REALLOCATED(int length_allocated, int size);
 typedef struct __st_array_list {
     // public
     ArrayListType type; // Tipo de dados armazenado pelo array
-    int size; // tamanho so array
-    int min_extra; // Quantidade mínima de espaço extra na realocação do Array
+    size_t size; // tamanho so array
+    size_t min_extra; // Quantidade mínima de espaço extra na realocação do Array
     ArrayListReallocateStrategy* reallocate_strategy; // Estratégia de realocação de espaço
 
     // private
     void* __pointer; // Ponteiro que armazena os alementos do array
-    int __length_allocated; // Espaço alocado na memória
+    size_t __length_allocated; // Espaço alocado na memória
 } ArrayList;
 
 /**
@@ -108,7 +108,7 @@ void* array_list_pointer(ArrayList* array_list);
  * @param min_extra valor extra mínimo na realocação do ArrayList
  * @param reallocate_strategy estratégia para realocação
  */
-void array_list_init_reallocate_strategy(ArrayList* array_list, ArrayListType type, int min_length_allocated, int min_extra, ArrayListReallocateStrategy* reallocate_strategy);
+void array_list_init_reallocate_strategy(ArrayList* array_list, ArrayListType type, size_t min_length_allocated, size_t min_extra, ArrayListReallocateStrategy* reallocate_strategy);
 
 /**
  * Inicializa um ArrayList vazio
@@ -120,7 +120,7 @@ void array_list_init_reallocate_strategy(ArrayList* array_list, ArrayListType ty
  * @param type tipo de dados armazenado pelo ArrayList
  * @param min_length_allocated quantidade mínima que deve estar alocado
  */
-void array_list_init_allocated(ArrayList* array_list, ArrayListType type, int min_length_allocated);
+void array_list_init_allocated(ArrayList* array_list, ArrayListType type, size_t min_length_allocated);
 
 /**
  * Inicializa um ArrayList vazio
@@ -145,7 +145,7 @@ void array_list_init(ArrayList* array_list, ArrayListType type);
  * @param reallocate_strategy estratégia para realocação
  * @return nova instância de ArrayList
  */
-ArrayList* new_array_list_reallocate_strategy(ArrayListType type, int min_length_allocated, int min_extra, ArrayListReallocateStrategy* reallocate_strategy);
+ArrayList* new_array_list_reallocate_strategy(ArrayListType type, size_t min_length_allocated, size_t min_extra, ArrayListReallocateStrategy* reallocate_strategy);
 
 /**
  * Cria um novo ArrayList vazio
@@ -157,7 +157,7 @@ ArrayList* new_array_list_reallocate_strategy(ArrayListType type, int min_length
  * @param min_length_allocated quantidade mínima que deve estar alocado
  * @return nova instância de ArrayList
  */
-ArrayList* new_array_list_allocated(ArrayListType type, int min_length_allocated);
+ArrayList* new_array_list_allocated(ArrayListType type, size_t min_length_allocated);
 
 /**
  * Cria um novo ArrayList vazio
@@ -174,7 +174,7 @@ ArrayList* new_array_list(ArrayListType type);
  * @param array_list instância do ArrayList
  * @return quantidade de espaço alocado para o ArrayList
  */
-int array_list_get_length_allocated(ArrayList* array_list);
+size_t array_list_get_length_allocated(ArrayList* array_list);
 
 /**
  * Informa a quantidade mínima de espaço que deve estar alocado
@@ -185,7 +185,7 @@ int array_list_get_length_allocated(ArrayList* array_list);
  * @param length_allocated quantidade mínima que deve estar alocada
  * @return 0 se uma realocação foi necessária
  */
-short array_list_set_min_length_allocated(ArrayList* array_list, int length_allocated);
+short array_list_set_min_length_allocated(ArrayList* array_list, size_t length_allocated);
 
 /**
  * Informa a quantidade máxima de espaço que deve estar alocado
@@ -196,7 +196,7 @@ short array_list_set_min_length_allocated(ArrayList* array_list, int length_allo
  * @param length_allocated quantidade máxima que deve estar alocada
  * @return 0 se uma realocação foi necessária
  */
-short array_list_set_max_length_allocated(ArrayList* array_list, int length_allocated);
+short array_list_set_max_length_allocated(ArrayList* array_list, size_t length_allocated);
 
 /**
  * Informa a quantidade exata que deve estar alocado para um determinado
@@ -208,7 +208,7 @@ short array_list_set_max_length_allocated(ArrayList* array_list, int length_allo
  * @param length_allocated quantidade exata que deve estar alocada
  * @return 0 se uma realocação ocorreu
  */
-short array_list_set_length_allocated(ArrayList* array_list, int length_allocated);
+short array_list_set_length_allocated(ArrayList* array_list, size_t length_allocated);
 
 /**
  * Adiciona no final do ArrayList os valores presentes em "values"
@@ -217,7 +217,7 @@ short array_list_set_length_allocated(ArrayList* array_list, int length_allocate
  * @param values array contendo os valores a serem atribuídos
  * @param size quantidade de valores a serem adicionados
  */
-void array_list_add_all(ArrayList* array_list, void* values, int size);
+void array_list_add_all(ArrayList* array_list, void* values, size_t size);
 
 /**
  * Adiciona em uma posição específica do ArrayList os valores presentes em "values"
@@ -227,7 +227,7 @@ void array_list_add_all(ArrayList* array_list, void* values, int size);
  * @param size quantidade de valores a serem adicionados
  * @param index posição a ser adicionado os valores
  */
-void array_list_add_all_at(ArrayList* array_list, void* values, int size, int index);
+void array_list_add_all_at(ArrayList* array_list, void* values, size_t size, size_t index);
 
 /**
  * Ordena o ArrayList
@@ -237,7 +237,7 @@ void array_list_add_all_at(ArrayList* array_list, void* values, int size, int in
  * @param begin primeira posição a ser ordenada (inclusive)
  * @param end última posição a ser ordenada (exclusive)
  */
-void array_list_sort(ArrayList* array_list, int (*cmp)(void*, void*), int begin, int end);
+void array_list_sort(ArrayList* array_list, size_t (*cmp)(void*, void*), size_t begin, size_t end);
 
 // ### implements interface_list.h ###
 
@@ -248,7 +248,7 @@ void array_list_sort(ArrayList* array_list, int (*cmp)(void*, void*), int begin,
  * @param index posição do alemento
  * @return a referência do valor buscado. NULL caso não seja encontrado
  */
-void* array_list_get_at(ArrayList* array_list, int index);
+void* array_list_get_at(ArrayList* array_list, size_t index);
 
 /**
  * Atualiza o valor do elemento do ArrayList que possui a posição fornecida
@@ -257,7 +257,7 @@ void* array_list_get_at(ArrayList* array_list, int index);
  * @param index posição do elemento a ser atualizado
  * @param value valor a ser atribuído
  */
-void array_list_set_at(ArrayList* array_list, int index, void* value);
+void array_list_set_at(ArrayList* array_list, size_t index, void* value);
 
 /**
  * Busca a posição no ArrayList da referência fornecida
@@ -266,7 +266,7 @@ void array_list_set_at(ArrayList* array_list, int index, void* value);
  * @param value referência a ser buscada
  * @return posição na lista da referência. -1 se não for encontrada
  */
-int array_list_find_index_by_reference(ArrayList* array_list, void* value);
+size_t array_list_find_index_by_reference(ArrayList* array_list, void* value);
 
 /**
  * Adiciona no final do ArrayList um valor
@@ -283,7 +283,7 @@ void array_list_add(ArrayList* array_list, void* value);
  * @param value valor a ser adicionado
  * @param index posição a ser adicionada o valor
  */
-void array_list_add_at(ArrayList* array_list, void* value, int index);
+void array_list_add_at(ArrayList* array_list, void* value, size_t index);
 
 /**
  * Remove do ArrayList e apaga da memoria a refêrencia fornecida
@@ -302,7 +302,7 @@ short array_list_eraser_by_reference(ArrayList* array_list, void* value);
  * @param index posição a ter o elemento removido do ArrayList e apagado
  *        da memória
  */
-void array_list_eraser_at(ArrayList* array_list, int index);
+void array_list_eraser_at(ArrayList* array_list, size_t index);
 
 /**
  * Remove do ArrayList sem apagar da memoria a refêrencia fornecida
@@ -322,7 +322,7 @@ short array_list_remove_by_reference(ArrayList* array_list, void* value);
  *        apagado da memória
  * @return referência removida do ArrayList
  */
-void* array_list_remove_at(ArrayList* array_list, int index);
+void* array_list_remove_at(ArrayList* array_list, size_t index);
 
 /**
  * Remove os elementos do ArrayList sem apagar os elementos da memoria.
@@ -382,6 +382,6 @@ void array_list_free_eraser_destructor(ArrayList* array_list, void (*destructor)
  *        1 para continuar a iteração. 0 para que o comando
  *        break seja executado
  */
-void array_list_for_each(ArrayList* array_list, short (*callback)(void*, int));
+void array_list_for_each(ArrayList* array_list, short (*callback)(void*, size_t));
 
 #endif // C_EMD_UTIL_ARRAY_LIST_H_INCLUDED
